fix(intr): return the whole interrupt vector, and say when there is none
intr_request_c holds its vector in a uint16_t and the device parameter that
sets it is nine bits wide - a PDP-11 vector is a byte address in the trap page
and the floating vectors of a second controller live above 0400 - but
get_vector() returned uint8_t, so bit 8 was dropped on the way out. The one
caller is the resource line of get_qunibus_resource_info(), listed by the
device menu's "ld", which therefore showed a wrong octal vector for anything
from 0400 up.

Widening the getter uncovered what the truncation had been hiding: the
constructor sets vector = 0xffff for "not set yet", and an MSCP port has no
vector until the guest's init handshake gives it one, so uda and tqk50 were
listing a plausible-looking 377 that was nothing of the sort. The line now
prints "5/-" for a request with no vector rather than a number nobody assigned.

Verified on ubx with the device menu, service stopped: with rl's intr_vector
set to 0500, the installed binary printed "INTRs:5/100" and the new one prints
"INTRs:5/500". Vectors below 0400 are unchanged (ts 5/224, dzv11 4/300,4/304,
DL11 4/060,4/064), and uda and tqk50 went from "5/377" to "5/-". Board
restored to the xxx configuration afterwards: XXDP boots from DL0 and lists
its directory, and rl's vector is back at its 0160 default.

+++ 10.01_base/2_src/arm/priorityrequest.hpp
@@ -173,7 +173,10 @@ public:
 
 	void set_level(uint8_t level);
 	void set_vector(uint16_t vector);
-	uint8_t get_vector(void) {
+	// uint16_t, like the vector itself: a PDP-11 vector is a 9 bit byte address
+	// into the trap page, and the floating vectors of a second controller live
+	// above 0400. Returning uint8_t silently dropped that bit.
+	uint16_t get_vector(void) {
 		return vector;
 	}
